ESlint fix and project restructured
